Add CNI test to integration suite (#20429)
* Add CNI test to integration suite

For: https://github.com/istio/istio/issues/20335

Right now we have a cni-simpleTest. We want to remove simpleTest, so we
need CNI coverage in integration/.

This test is just a simple re-use of the existing reachability tests
that sanity checks that we get all iptables stuff set up properly and
nothing blows up (which is more than the previous test covered).

* add license

* Set hub and tag

* workaround https://github.com/istio/istio/issues/20439