Fix finalization in operator (#20205)
* Properly handle conflicts when adding finalizer

* Properly handle conflicts when removing finalizer

* Retry finalizer removal only once

Only the first retry has a real chance of succeeding, due to the relatively long
time between the initial get and the update (reconciler.Delete() takes several seconds).
If the object is updated during this time, there's a high chance the cache (which
the get operation is retrieving the object from) will be updated, which will
allow the retry to succeed. But if the cache is not updated, retrying an
additional 9 times is almost certainly bound to fail due to the low probability
of the cache being updated between two retries.

* Remove finalizer only if cleanup was successful