Support standalone Galley for MCP (#20517)
Right now, to run Galley, you need certs (and therefor citadel) and it
will try to do validation which won't work out because it will fight
with istiod validation.

This change makes it so we only require certs if validation is enabled
and istiod is disabled. We also fix liveness probes in galley to support
this, and remove places where we deployed citadel just for galley.