Drop all capabilities from control-plane containers (#20854)
* Drop all capabilities from control-plane containers

As suggested during discussion of #20311. These containers don't run as root
and don't need any capabilities enabled.

* Set workdir to / for distroless-based images

The distroless base has workdir as /home/nonroot but we run as 1337:1337
which has no permissions on that directory so the runtime fails to start
the container. The standard base has workdir / so change to match that.

* Remove duplicate WORKDIR directive