Fix validation webhook on older k8s (#21242)
On older versions, the port field is not supported. It defaults to 443,
which we select anyways, so we should just remove the default.

This is a regression caused by #21164 - its not clear to me why it
worked before though. I guess when galley created the config it wasn't
validated or something maybe?