Update zipkin deployment (#21686)
* Update zipkin deployment

The image used is quite old. In updating, I found a number of issues
impacting startup speed. We have a delay of 200s to the readiness probe,
so we will never startup in less than 200s. Additionally, because the
CPU limit is so small, startup takes 50s, vs 5s now that the limit is
bumped up a little bit. Memory is also raised, as this is a JVM
appliction so it uses a lot of memory, and has often been OOMKilled in
my tests.

Additionally, I cleaned up a TODO in the tests so that the correct
OWNERS get assigned here.

* gen