Stop skipping verification of TLS in tests (#22177)
This was added recently in
https://github.com/istio/istio/pull/22052 which was skipping
verification.

This PR:
* Adds back verification
* Fixes to properly set the ca cert, to allow verification to work
* Fixes to allow overriding hostname, to allow verification of DNS certs
to work
* Adds a new DNS based cert to the standard certs folder
* Starts using the standard certs, remove the custom ones I had added
just for this test. This required making a new DNS based cert though, as
the other one is a SPIFFEE cert which does not seem easily verifiable
without using custom SPIFFEE libraries. It is also not really what we
want to test, as this is for custom TLS, not Istio TLS, so users are not
just using SPIFFEE.