Warn on missing IstioOperator file in DefaultConfig (#22520)
This changes `DefaultConfig` in the test framework to only print a
warning when the default `IstioOperator` file can't be found rather
than return an error.  Otherwise, it may be impossible to override the
path in a user supplied setup function because `DefaultConfig` is
called before the setup function, and if it errors, setup is aborted
early.