Fix prometheus operator for istiod (#22529)
A recent change made our grafana dashboards query based on the app label
instead of the job label. The operator configs do not have the app
label. This change adds the app label, and additionally adds istio=pilot
label to the Service for consistency (otherwise we will need another
ServiceMonitor)