Support TLS on all protocols in echo (#22656)
* Support TLS on all protocols in echo

Currently only gRPC respects the TLS settings, meaning we cannot test
HTTP, or TCP over TLS. This extends echo to take support these
protocols. Additionally, I extended one of the tests that uses this to
test TCP and HTTP, which validates this works correctly.

I think it makes sense for a single echo to be able to serve TLS and non
TLS traffic as well, so I moved the TLS configuration to the port level.
TLS ports are specified with --tls, and in the echo config `TLS: true`.

There are some other tests that can be extended to include these
changes, but for now I will keep this PR smaller.

* fix lint