Only apply latest EnvoyFilters when revision is set (#22877)
Currently there are issues when applying a new 1.6 revision with an old
control plane running. For example, 1.4+mixer. When 1.6 revision is
added, we will create a bunch of EnvoyFilters. These are supposed to be
scoped to just that revision (by istio.io/rev label), but the old 1.4
control plane doesn't understand this label. This means we will get
v2+mixer and get double telemetry, and in general even if it wasn't
double telemetry we don't want our canary to impact the stable version.

There are two possible fixes for this:
1 Backport istio.io/rev support to 1.4 or even older, so they can update
to latest patch release.
2 Only install the latest EnvoyFilters when a revision is set. We will
never have a 1.5 proxy connect to a 1.6 revision, so this is never
losing config we expect. This way the EnvoyFilter will not apply to old
1.4/1.5 proxies. For 1.6 proxies in default revision this will also not
match, as the 1.6 Pilot respoects the istio.io/rev label

This PR implements the second option. We may want to also implement (1)
for other reasons, but this is useful regardless since it will work with
any release of Istio